Make mineral P rates agree with Eurostat GNBs
In particular this makes a large difference for Italy and Cyprus.

At least for Italy, the aei_fm_usefert data seems to be wrong;
the aei_fm_manfert table agrees better with national statistics
for Italy.

Eurostat's GNB calculations appear to use a mix of usefert and
manfert values, and in some cases something else.